+++ src/main/java/com/codegym/controller/AuthController.java
@@ -38,42 +38,10 @@ public class AuthController {
     private AuthenticationManager authenticationManager;
 
     @Autowired
-    private IUserService userService;
-
-    @Autowired
-    private IUserRepository userRepository;
-
-    @Autowired
-    private RoleRepository roleRepository;
-
-    @Autowired
     PasswordEncoder passwordEncoder;
 
     @Autowired
     JwtTokenProvider tokenProvider;
-
-
-    @PostMapping("/register")
-    public ResponseEntity<?> registerUser(@Valid @RequestBody RegisterRequest registerRequest) {
-        if(userRepository.existsByUsername(registerRequest.getUsername())) {
-            return new ResponseEntity<>(new RegisterResponse("Username is already taken!"), HttpStatus.BAD_REQUEST);
-        }
-
-        if(userRepository.existsByEmail(registerRequest.getEmail())) {
-            return new ResponseEntity<>(new RegisterResponse("Email Address already in use!"), HttpStatus.BAD_REQUEST);
-        }
-
-        User user = new User(registerRequest.getUsername(),
-                registerRequest.getEmail(),
-                passwordEncoder.encode(registerRequest.getPassword()));
-
-        Role userRole = roleRepository.findByName("ROLE_USER").orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-        user.setRoles(Collections.singleton(userRole));
-
-        userService.save(user);
-
-        return ResponseEntity.ok(new RegisterResponse("User registered successfully"));
-    }
 
 
     @PostMapping("/login")
@@ -86,7 +54,7 @@ public class AuthController {
 
             SecurityContextHolder.getContext().setAuthentication(authentication);
 
-            // Gọi hàm tạo Token
+
             String token = tokenProvider.generateToken(authentication);
             return new ResponseEntity<>(new LoginResponse("Login success!", token), HttpStatus.OK);
         } catch (Exception e) {
