+++ src/main/java/com/codegym/controller/UserController.java
@@ -1,4 +1,62 @@
 package com.codegym.controller;
 
+import com.codegym.model.Role;
+import com.codegym.model.User;
+import com.codegym.payload.request.RegisterRequest;
+import com.codegym.payload.response.RegisterResponse;
+import com.codegym.repository.IUserRepository;
+import com.codegym.repository.RoleRepository;
+import com.codegym.security.JwtTokenProvider;
+import com.codegym.service.IUserService;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Collections;
+
+
+@CrossOrigin(value = "*", maxAge = 3600)
+@RestController
+@RequestMapping("/api/auth")
 public class UserController {
+    @Autowired
+    private IUserService userService;
+
+    @Autowired
+    private IUserRepository userRepository;
+
+    @Autowired
+    private RoleRepository roleRepository;
+
+    @Autowired
+    PasswordEncoder passwordEncoder;
+
+    @Autowired
+    JwtTokenProvider tokenProvider;
+
+
+    @PostMapping("/register")
+    public ResponseEntity<?> registerUser(@Valid @RequestBody RegisterRequest registerRequest) {
+        if(userRepository.existsByUsername(registerRequest.getUsername())) {
+            return new ResponseEntity<>(new RegisterResponse("Username is already taken!"), HttpStatus.BAD_REQUEST);
+        }
+
+        if(userRepository.existsByEmail(registerRequest.getEmail())) {
+            return new ResponseEntity<>(new RegisterResponse("Email Address already in use!"), HttpStatus.BAD_REQUEST);
+        }
+
+        User user = new User(registerRequest.getUsername(),
+                registerRequest.getEmail(),
+                passwordEncoder.encode(registerRequest.getPassword()));
+
+        Role userRole = roleRepository.findByName("ROLE_USER").orElseThrow(() -> new RuntimeException("Error: Role is not found."));
+        user.setRoles(Collections.singleton(userRole));
+
+        userService.save(user);
+
+        return ResponseEntity.ok(new RegisterResponse("User registered successfully"));
+    }
 }
